Separate setup policy from the rest of project planning
ProjectPlanning is a big module, counting more than 4k lines.

This change separates the code related to the setup policy to its own
module and tidies up a bit the documentation of ProjectPlanning.
